requirements.txt
Flask
requests
pyrebase
google-cloud-datastore==1.7.3